requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
billiard==4.2.2
build==1.3.0
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
chromadb==1.2.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
dataclasses-json==0.6.7
dill==0.4.0
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.7
django-debug-toolbar==6.1.0
django_celery_results==2.6.0
docling==2.58.0
docling-core==2.49.0
docling-ibm-models==3.10.2
docling-parse==4.7.0
docopt==0.6.2
durationpy==0.10
et_xmlfile==2.0.0
Faker==37.12.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
google-auth==2.41.1
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
kubernetes==34.1.0
langchain-chroma==1.0.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.2
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langsmith==0.4.39
latex2mathml==3.78.1
lxml==5.4.0
Markdown==3.10
markdown-it-py==4.0.0
marko==2.2.1
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpire==2.10.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pipreqs==0.4.13
pluggy==1.6.0
polyfactory==2.22.3
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pylatexenc==2.10
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
rapidocr==3.4.2
redis==7.0.1
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
semchunk==2.2.2
sentence-transformers==5.1.2
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlparse==0.5.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
whitenoise==6.11.0
xlsxwriter==3.2.9
yarg==0.1.10
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
python-Levenshtein==0.26.1
matplotlib==3.8.1
seaborn